Remove '-a' option.
Initially I had written gcov2html-check to have two modes: absolute
and non-absolute. In non-absolute mode the colourization would be
based on the number of runs, which assumes most lines of code are
only executed once or so per run. In absolute mode the entire file
is examined to find the line that is run the most often, then all
the colours are based on that max count.

I don't believe the non-absolute mode is useful, therefore the code
will always run in absolute mode. Therefore the switch to explicitly
define the state is no longer necessary.